# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IDEABOX SWAGGER",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/favourites": {
            "get": {
                "tags": [
                    "Favourites"
                ],
                "summary": "Get all favourites",
                "operationId": "c8a645915581acd65f95cf476eb7bf5a",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Favourites"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Favourites"
                ],
                "summary": "Create a new favourite",
                "operationId": "dce3e6af9abcf6bdb8c60f0288b8ac55",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "generate_date",
                                    "value",
                                    "name",
                                    "category_id"
                                ],
                                "properties": {
                                    "generate_date": {
                                        "type": "string",
                                        "example": "2024-05-05"
                                    },
                                    "value": {
                                        "type": "string",
                                        "example": "Some value"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Favourite Name"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Favourite created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Favourites"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/favourites/{id}": {
            "get": {
                "tags": [
                    "Favourites"
                ],
                "summary": "Get a specific favourite",
                "operationId": "079c6613d2611a39d514d44567a21e16",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Favourite ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Favourites"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Favourite not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Favourite not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Favourites"
                ],
                "summary": "Update a favourite",
                "operationId": "f2726b81bdfa6c8d1d450801a5cb5e2e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Favourite ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "generate_date": {
                                        "type": "string",
                                        "example": "2024-05-06"
                                    },
                                    "value": {
                                        "type": "string",
                                        "example": "Updated value"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Favourite Name"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Favourite updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Favourites"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Favourite not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Favourite not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Favourites"
                ],
                "summary": "Delete a favourite",
                "operationId": "36bbcae90b4a41357724abd080bfcf55",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Favourite ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Favourite deleted successfully"
                    },
                    "404": {
                        "description": "Favourite not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Favourite not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Favourites": {
                "properties": {
                    "id": {
                        "description": "The unique identifier for the favourite",
                        "type": "integer",
                        "format": "int64"
                    },
                    "generate_date": {
                        "description": "The date when the favourite was generated",
                        "type": "string",
                        "format": "date"
                    },
                    "value": {
                        "description": "The value of the favourite",
                        "type": "string"
                    },
                    "name": {
                        "description": "The name of the favourite",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "The ID of the user who created the favourite",
                        "type": "integer",
                        "format": "int64"
                    },
                    "category_id": {
                        "description": "The ID of the category to which the favourite belongs",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "The date and time when the favourite was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The date and time when the favourite was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}